requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
black==24.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
docstring-parser==0.15
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
grpcio==1.62.0
grpcio-tools==1.62.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
piccolo==1.3.0
platformdirs==4.2.0
protobuf==4.23.1
pydantic==2.6.1
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
targ==0.3.8
tomli==2.0.1
typing_extensions==4.9.0
opentelemetry-api==1.15.0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
